requirements.txt
django
psycopg2
uwsgi
requests
dj_config_url
djangorestframework
